Refactors react to separate classes
