+++ count_server/mix.exs
@@ -28,17 +28,9 @@ defmodule CountServer.Mixfile do
           :logger,
           :comeonin
         ]
-      ] #app_list(Mix.env)
+      ]
   end
 
-  # def app_list do
-  #   [:phoenix, :phoenix_html, :cowboy, :logger, :gettext, :phoenix_ecto, :postgrex]
-  # end
-  #
-  # def app_list(:test), do: [:hound | app_list]
-  # def app_list(_), do: app_list
-
-  # Specifies which paths to compile per environment.
   defp elixirc_paths(:test), do: ["lib", "web", "test/support"]
   defp elixirc_paths(_),     do: ["lib", "web"]
